Weapon article update (+ slight macro change).
I have changed the layout of the weapons article, which resulted in it:
- Being split into different sections, each one dedicated to specific weapon type, resulting in a more organised look;
- Added a a "contents" list for improved navigation;
- Added short lists of general characteristics for each weapon type.

To achieve this, however, I had to slightly change the code responsible for renderig the content tables - the change simply being the addition of an additional argument for the function rendering the content table, allowing it to generate a table containg only the specified weapon type.
I did read the documentation for Tera for Rust and the macro should be fine - I do not have the ability to test it on my end, unfortunately.